Move APIResponse to api.py and update MARC objects to use it
