+++ src/alma_api_client/__init__.py
@@ -1,3 +1,3 @@
-from .alma_api_client import AlmaAPIClient, APIResponse  # TODO: Remove APIResponse here
+from .alma_api_client import AlmaAPIClient
 from .alma_analytics_client import AlmaAnalyticsClient
 from .alma_marc import get_pymarc_record_from_bib, prepare_bib_for_update
